split submit screen in components
