+++ mobile/bilmenu/app/(tabs)/submit.tsx
@@ -1,41 +1,14 @@
 import React, { useState, useEffect } from "react";
-import {
-  StyleSheet,
-  View,
-  Text,
-  TouchableOpacity,
-  TextInput,
-  ScrollView,
-  Alert,
-  Image,
-  Switch,
-  KeyboardAvoidingView,
-  Platform,
-  Modal,
-} from "react-native";
+import { StyleSheet, View, Text, Alert } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { Header } from "@/components/header";
 import { BilMenuTheme } from "@/constants/theme";
 import * as ImagePicker from "expo-image-picker";
 import * as MailComposer from "expo-mail-composer";
-
-interface Meal {
-  tr: string;
-  en: string;
-}
-
-interface MealPlanDay {
-  date: string;
-  dayOfWeek: string;
-  lunch: Meal[];
-  dinner: Meal[];
-  alternative: Meal[];
-}
-
-interface MissingMeal {
-  name: string;
-  isMissing: boolean;
-}
+import { Step, MealType, Step1View } from "@/types/submit";
+import { getCurrentDayOfWeek, getCurrentMealTime } from "@/utils/submitUtils";
+import { useSubmitData } from "@/hooks/useSubmitData";
+import { Step1, Step2, Step3 } from "@/components/submit";
 
 export default function SubmitScreen() {
   const [selectedMeal, setSelectedMeal] = useState<string>("");
@@ -44,87 +17,23 @@ export default function SubmitScreen() {
   const [userName, setUserName] = useState<string>("");
   const [showAttributionModal, setShowAttributionModal] =
     useState<boolean>(false);
-  const [currentWeekMeals, setCurrentWeekMeals] = useState<MissingMeal[]>([]);
   const [searchQuery, setSearchQuery] = useState("");
-  const [isLoading, setIsLoading] = useState(true);
-  const [step, setStep] = useState<1 | 2 | 3>(1);
+  const [step, setStep] = useState<Step>(1);
   const [showAllMissing, setShowAllMissing] = useState(false);
-  const [allMissingMeals, setAllMissingMeals] = useState<MissingMeal[]>([]);
-  const [lastUpdated, setLastUpdated] = useState<string>("");
-  const [totalCheckpoints, setTotalCheckpoints] = useState<number>(0);
-  const [step1View, setStep1View] = useState<"search" | "daymeal">("daymeal");
+  const [step1View, setStep1View] = useState<Step1View>("daymeal");
   const [selectedDay, setSelectedDay] = useState<number>(0);
-  const [selectedMealType, setSelectedMealType] = useState<
-    "lunch" | "dinner" | "alternative"
-  >("lunch");
-  const [currentWeekMealPlan, setCurrentWeekMealPlan] = useState<MealPlanDay[]>(
-    []
-  );
+  const [selectedMealType, setSelectedMealType] = useState<MealType>("lunch");
 
-  // Get current week number (ISO week)
-  const getCurrentWeek = () => {
-    const now = new Date();
-    const target = new Date(now.valueOf());
-    const dayNr = (now.getDay() + 6) % 7;
-    target.setDate(target.getDate() - dayNr + 3);
-    const firstThursday = target.valueOf();
-    target.setMonth(0, 1);
-    if (target.getDay() !== 4) {
-      target.setMonth(0, 1 + ((4 - target.getDay() + 7) % 7));
-    }
-    return 1 + Math.ceil((firstThursday - target.valueOf()) / 604800000);
-  };
-
-  // Get current day of week (Monday = 0, Sunday = 6)
-  const getCurrentDayOfWeek = () => {
-    const today = new Date();
-    const day = today.getDay();
-    return day === 0 ? 6 : day - 1; // Convert Sunday (0) to 6, others shift by 1
-  };
-
-  // Get current meal time based on hour
-  const getCurrentMealTime = (): "lunch" | "dinner" | "alternative" => {
-    const now = new Date();
-    const hour = now.getHours();
-
-    // Lunch: 11:00 - 15:00
-    if (hour >= 11 && hour < 15) {
-      return "lunch";
-    }
-    // Dinner: 17:00 - 21:00
-    else if (hour >= 17 && hour < 21) {
-      return "dinner";
-    }
-    // Default to lunch for other times
-    else {
-      return "lunch";
-    }
-  };
-
-  // Get day name
-  const getDayName = (dayIndex: number) => {
-    const days = [
-      "Monday",
-      "Tuesday",
-      "Wednesday",
-      "Thursday",
-      "Friday",
-      "Saturday",
-      "Sunday",
-    ];
-    return days[dayIndex];
-  };
-
-  // Get short day name
-  const getShortDayName = (dayIndex: number) => {
-    const days = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"];
-    return days[dayIndex];
-  };
-
-  // Load missing meals data
-  useEffect(() => {
-    loadMissingMeals();
-  }, []);
+  // Use the custom hook for data management
+  const {
+    currentWeekMeals,
+    allMissingMeals,
+    currentWeekMealPlan,
+    isLoading,
+    lastUpdated,
+    totalCheckpoints,
+    getMissingMealsForDayAndType,
+  } = useSubmitData();
 
   // Auto-select current day and meal time
   useEffect(() => {
@@ -134,126 +43,16 @@ export default function SubmitScreen() {
     setSelectedMealType(currentMealTime);
   }, []);
 
-  const loadMissingMeals = async () => {
-    try {
-      setIsLoading(true);
-
-      // Get current week and year
-      const now = new Date();
-      const currentWeek = getCurrentWeek();
-      const currentYear = now.getFullYear();
-
-      // First, get the latest checkpoint URL from the state file
-      const latestCheckpointResponse = await fetch(
-        "https://raw.githubusercontent.com/ndricimrr/bilmenu/refs/heads/main/webapp/missing-meals/latest-checkpoint.json"
-      );
-
-      if (!latestCheckpointResponse.ok) {
-        throw new Error("Failed to load latest checkpoint state");
-      }
-
-      const latestCheckpointData = await latestCheckpointResponse.json();
-      const latestCheckpointUrl = latestCheckpointData.latestCheckpointUrl;
-
-      // Store metadata for display
-      setLastUpdated(latestCheckpointData.lastUpdated);
-      setTotalCheckpoints(latestCheckpointData.totalCheckpoints);
-
-      // Load the latest missing meals checkpoint using the dynamic URL
-      const missingMealsResponse = await fetch(latestCheckpointUrl);
-
-      if (!missingMealsResponse.ok) {
-        throw new Error("Failed to load missing meals data");
-      }
-
-      const missingMealsData = await missingMealsResponse.json();
-      const missingMealsSet = new Set(missingMealsData.missingMeals);
-
-      // Try to load current week's meal plan
-      let currentWeekMeals: MissingMeal[] = [];
-
-      try {
-        const mealPlanResponse = await fetch(
-          `https://raw.githubusercontent.com/ndricimrr/bilmenu/refs/heads/main/webapp/mealplans/meal_plan_week_${currentWeek}_${currentYear}.json`
-        );
-
-        if (mealPlanResponse.ok) {
-          const mealPlanData = await mealPlanResponse.json();
-          const allMeals = new Set<string>();
-
-          // Store the meal plan data for day/meal view
-          if (Array.isArray(mealPlanData)) {
-            setCurrentWeekMealPlan(mealPlanData);
-
-            mealPlanData.forEach((day: MealPlanDay) => {
-              if (day.lunch)
-                day.lunch.forEach((meal: Meal) => allMeals.add(meal.tr));
-              if (day.dinner)
-                day.dinner.forEach((meal: Meal) => allMeals.add(meal.tr));
-              if (day.alternative)
-                day.alternative.forEach((meal: Meal) => allMeals.add(meal.tr));
-            });
-          }
-
-          // Filter to only show missing meals from current week
-          currentWeekMeals = Array.from(allMeals)
-            .filter((meal) => missingMealsSet.has(meal))
-            .map((meal) => ({ name: meal, isMissing: true }));
-        }
-      } catch (mealPlanError) {
-        console.log(
-          "Could not load current week meal plan, showing all missing meals"
-        );
-      }
-
-      // Store all missing meals
-      const allMissing = missingMealsData.missingMeals.map((meal: string) => ({
-        name: meal,
-        isMissing: true,
-      }));
-      setAllMissingMeals(allMissing);
-
-      // If no current week meals found, show all missing meals
-      if (currentWeekMeals.length === 0) {
-        setCurrentWeekMeals(allMissing);
-      } else {
-        setCurrentWeekMeals(currentWeekMeals);
-      }
-    } catch (error) {
-      console.error("Error loading missing meals:", error);
-      Alert.alert(
-        "Error",
-        "Failed to load missing meals data. Please try again later."
-      );
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  // Get missing meals for selected day and meal type
-  const getMissingMealsForDayAndType = () => {
-    if (currentWeekMealPlan.length === 0 || !allMissingMeals.length) {
-      return [];
-    }
-
-    const dayData = currentWeekMealPlan[selectedDay];
-    if (!dayData) return [];
-
-    const mealsForType = dayData[selectedMealType] || [];
-    const missingMealsSet = new Set(allMissingMeals.map((meal) => meal.name));
-
-    return mealsForType
-      .filter((meal: Meal) => missingMealsSet.has(meal.tr))
-      .map((meal: Meal) => ({ name: meal.tr, isMissing: true }));
-  };
-
   const mealsToShow = showAllMissing ? allMissingMeals : currentWeekMeals;
   const filteredMeals = mealsToShow.filter((meal) =>
     meal.name.toLowerCase().includes(searchQuery.toLowerCase())
   );
 
   // Get meals for day/meal view
-  const dayMealViewMeals = getMissingMealsForDayAndType();
+  const dayMealViewMeals = getMissingMealsForDayAndType(
+    selectedDay,
+    selectedMealType
+  );
 
   const handleMealSelection = (mealName: string) => {
     setSelectedMeal(mealName);
@@ -371,441 +170,47 @@ ${userName || "A Bilkent student"}`;
   };
 
   const renderStep1 = () => (
-    <ScrollView
-      style={styles.stepContainer}
-      showsVerticalScrollIndicator={true}
-    >
-      <Text style={styles.stepTitle}>Step 1: Select Missing Meal</Text>
-
-      <View style={styles.weekInfo}>
-        <Text style={styles.weekInfoText}>
-          Week {getCurrentWeek()}, {new Date().getFullYear()}
-        </Text>
-        {lastUpdated && (
-          <Text style={styles.lastUpdatedText}>
-            Data updated: {new Date(lastUpdated).toLocaleDateString()}
-          </Text>
-        )}
-        {totalCheckpoints > 0 && (
-          <Text style={styles.checkpointCountText}>
-            {totalCheckpoints} checkpoint{totalCheckpoints !== 1 ? "s" : ""}{" "}
-            available
-          </Text>
-        )}
-      </View>
-
-      {/* View Selection Buttons */}
-      <View style={styles.viewSelector}>
-        <TouchableOpacity
-          style={[
-            styles.viewButton,
-            step1View === "daymeal" && styles.viewButtonActive,
-          ]}
-          onPress={() => setStep1View("daymeal")}
-        >
-          <Text
-            style={[
-              styles.viewButtonText,
-              step1View === "daymeal" && styles.viewButtonTextActive,
-            ]}
-          >
-            📅 Day & Meal
-          </Text>
-        </TouchableOpacity>
-        <TouchableOpacity
-          style={[
-            styles.viewButton,
-            step1View === "search" && styles.viewButtonActive,
-          ]}
-          onPress={() => setStep1View("search")}
-        >
-          <Text
-            style={[
-              styles.viewButtonText,
-              step1View === "search" && styles.viewButtonTextActive,
-            ]}
-          >
-            🔍 Search All
-          </Text>
-        </TouchableOpacity>
-      </View>
-
-      {step1View === "daymeal" ? renderDayMealView() : renderSearchView()}
-    </ScrollView>
-  );
-
-  const renderDayMealView = () => (
-    <>
-      {/* Day Selection */}
-      <View style={styles.daysContainer}>
-        <Text style={styles.sectionTitle}>Select Day</Text>
-        <View style={styles.daysGrid}>
-          {[0, 1, 2, 3, 4, 5, 6].map((dayIndex) => (
-            <TouchableOpacity
-              key={dayIndex}
-              style={[
-                styles.dayButton,
-                selectedDay === dayIndex && styles.dayButtonActive,
-                dayIndex === getCurrentDayOfWeek() && styles.dayButtonToday,
-              ]}
-              onPress={() => setSelectedDay(dayIndex)}
-            >
-              <Text
-                style={[
-                  styles.dayButtonText,
-                  selectedDay === dayIndex && styles.dayButtonTextActive,
-                ]}
-              >
-                {getShortDayName(dayIndex)}
-              </Text>
-              {dayIndex === getCurrentDayOfWeek() && (
-                <Text
-                  style={[
-                    styles.todayLabel,
-                    selectedDay === dayIndex && styles.todayLabelActive,
-                  ]}
-                >
-                  Today
-                </Text>
-              )}
-            </TouchableOpacity>
-          ))}
-        </View>
-      </View>
-
-      {/* Meal Type Selection */}
-      <View style={styles.mealTypeContainer}>
-        <Text style={styles.sectionTitle}>Select Meal Type</Text>
-        <View style={styles.mealTypeButtons}>
-          <TouchableOpacity
-            style={[
-              styles.mealTypeButton,
-              styles.mealTypeButtonLeft,
-              selectedMealType === "lunch" && styles.mealTypeButtonActive,
-            ]}
-            onPress={() => setSelectedMealType("lunch")}
-          >
-            <Text
-              style={[
-                styles.mealTypeButtonText,
-                selectedMealType === "lunch" && styles.mealTypeButtonTextActive,
-              ]}
-            >
-              ☀️ Lunch
-            </Text>
-          </TouchableOpacity>
-          <TouchableOpacity
-            style={[
-              styles.mealTypeButton,
-              styles.mealTypeButtonCenter,
-              selectedMealType === "dinner" && styles.mealTypeButtonActive,
-            ]}
-            onPress={() => setSelectedMealType("dinner")}
-          >
-            <Text
-              style={[
-                styles.mealTypeButtonText,
-                selectedMealType === "dinner" &&
-                  styles.mealTypeButtonTextActive,
-              ]}
-            >
-              🌙 Dinner
-            </Text>
-          </TouchableOpacity>
-          <TouchableOpacity
-            style={[
-              styles.mealTypeButton,
-              styles.mealTypeButtonRight,
-              selectedMealType === "alternative" && styles.mealTypeButtonActive,
-            ]}
-            onPress={() => setSelectedMealType("alternative")}
-          >
-            <Text
-              style={[
-                styles.mealTypeButtonText,
-                selectedMealType === "alternative" &&
-                  styles.mealTypeButtonTextActive,
-              ]}
-            >
-              💰 Alternative
-            </Text>
-          </TouchableOpacity>
-        </View>
-      </View>
-
-      {/* Missing Meals List */}
-      <ScrollView style={styles.mealsList} showsVerticalScrollIndicator={true}>
-        {isLoading ? (
-          <Text style={styles.loadingText}>Loading missing meals...</Text>
-        ) : dayMealViewMeals.length === 0 ? (
-          <Text style={styles.noResultsText}>
-            No missing meals found for {getDayName(selectedDay)}{" "}
-            {selectedMealType}.
-          </Text>
-        ) : (
-          <>
-            <Text style={styles.resultsCount}>
-              {dayMealViewMeals.length} missing meal
-              {dayMealViewMeals.length !== 1 ? "s" : ""} found for{" "}
-              {getDayName(selectedDay)} {selectedMealType}
-            </Text>
-            {dayMealViewMeals.map((meal, index) => (
-              <TouchableOpacity
-                key={index}
-                style={styles.mealButton}
-                onPress={() => handleMealSelection(meal.name)}
-              >
-                <Text style={styles.mealButtonIcon}>+</Text>
-                <Text style={styles.mealButtonText}>{meal.name}</Text>
-              </TouchableOpacity>
-            ))}
-          </>
-        )}
-      </ScrollView>
-    </>
-  );
-
-  const renderSearchView = () => (
-    <>
-      <View style={styles.toggleContainer}>
-        <View style={styles.toggleRow}>
-          <View style={styles.toggleInfo}>
-            <Text style={styles.toggleLabel}>
-              {showAllMissing ? "All Missing Meals" : "Current Week Only"}
-            </Text>
-          </View>
-          <Switch
-            value={showAllMissing}
-            onValueChange={setShowAllMissing}
-            trackColor={{
-              false: BilMenuTheme.colors.switchTrack,
-              true: BilMenuTheme.colors.secondary,
-            }}
-            thumbColor={
-              showAllMissing
-                ? BilMenuTheme.colors.switchThumbActive
-                : BilMenuTheme.colors.switchThumb
-            }
-          />
-        </View>
-      </View>
-
-      <TextInput
-        style={styles.searchInput}
-        placeholder="Search for a meal..."
-        value={searchQuery}
-        onChangeText={setSearchQuery}
-        placeholderTextColor={BilMenuTheme.colors.textLight}
-      />
-
-      <ScrollView style={styles.mealsList} showsVerticalScrollIndicator={true}>
-        {isLoading ? (
-          <Text style={styles.loadingText}>Loading missing meals...</Text>
-        ) : filteredMeals.length === 0 ? (
-          <Text style={styles.noResultsText}>
-            {searchQuery
-              ? "No meals found matching your search."
-              : "No missing meals found."}
-          </Text>
-        ) : (
-          <>
-            <Text style={styles.resultsCount}>
-              {filteredMeals.length} missing meal
-              {filteredMeals.length !== 1 ? "s" : ""} found
-            </Text>
-            {filteredMeals.map((meal, index) => (
-              <TouchableOpacity
-                key={index}
-                style={styles.mealButton}
-                onPress={() => handleMealSelection(meal.name)}
-              >
-                <Text style={styles.mealButtonIcon}>+</Text>
-                <Text style={styles.mealButtonText}>{meal.name}</Text>
-              </TouchableOpacity>
-            ))}
-          </>
-        )}
-      </ScrollView>
-    </>
+    <Step1
+      step1View={step1View}
+      onViewChange={setStep1View}
+      selectedDay={selectedDay}
+      selectedMealType={selectedMealType}
+      onDayChange={setSelectedDay}
+      onMealTypeChange={setSelectedMealType}
+      dayMealViewMeals={dayMealViewMeals}
+      searchQuery={searchQuery}
+      onSearchChange={setSearchQuery}
+      showAllMissing={showAllMissing}
+      onToggleShowAll={setShowAllMissing}
+      filteredMeals={filteredMeals}
+      isLoading={isLoading}
+      onMealSelect={handleMealSelection}
+      lastUpdated={lastUpdated}
+      totalCheckpoints={totalCheckpoints}
+    />
   );
 
   const renderStep2 = () => (
-    <View style={styles.stepContainer}>
-      <View style={styles.step2Content}>
-        <Text style={styles.stepTitle}>Step 2: Capture Image</Text>
-        <Text style={styles.selectedMealText}>
-          <Text style={styles.selectedLabel}>Selected: </Text>
-          <Text style={styles.selectedMealName}>{selectedMeal}</Text>
-        </Text>
-
-        <TouchableOpacity
-          style={styles.captureButton}
-          onPress={handleCameraCapture}
-        >
-          <Text style={styles.captureButtonText}>📷 Capture Image</Text>
-        </TouchableOpacity>
-
-        <TouchableOpacity
-          style={styles.galleryButton}
-          onPress={handleGallerySelection}
-        >
-          <Text style={styles.galleryButtonText}>🖼️ Select from Gallery</Text>
-        </TouchableOpacity>
-      </View>
-
-      <TouchableOpacity style={styles.backButton} onPress={() => setStep(1)}>
-        <Text style={styles.backButtonText}>← Back to Meal Selection</Text>
-      </TouchableOpacity>
-    </View>
+    <Step2
+      selectedMeal={selectedMeal}
+      onCameraCapture={handleCameraCapture}
+      onGallerySelection={handleGallerySelection}
+      onBack={() => setStep(1)}
+    />
   );
 
   const renderStep3 = () => (
-    <KeyboardAvoidingView
-      style={styles.stepContainer}
-      behavior={Platform.OS === "ios" ? "padding" : "height"}
-      keyboardVerticalOffset={Platform.OS === "ios" ? 50 : 50}
-      enabled={true}
-    >
-      <ScrollView
-        style={styles.step3Content}
-        showsVerticalScrollIndicator={false}
-        keyboardShouldPersistTaps="handled"
-        contentContainerStyle={styles.step3ScrollContent}
-        removeClippedSubviews={false}
-        scrollEventThrottle={16}
-      >
-        <Text style={styles.stepTitle}>Step 3: Send Email</Text>
-        <Text style={styles.selectedMealText}>
-          <Text style={styles.selectedLabel}>Meal: </Text>
-          <Text style={styles.selectedMealName}>{selectedMeal}</Text>
-        </Text>
-        <View style={styles.previewContainer}>
-          {capturedImage && (
-            <View style={styles.imagePreviewContainer}>
-              <View style={styles.imagePreviewSmall}>
-                <Image
-                  source={{ uri: capturedImage }}
-                  style={styles.previewImageSmall}
-                />
-              </View>
-              <Text style={styles.imageFilenameText}>{selectedMeal}.jpg</Text>
-              {imageSize && (
-                <Text style={styles.imageSizeText}>
-                  {(imageSize / (1024 * 1024)).toFixed(1)} MB
-                </Text>
-              )}
-            </View>
-          )}
-
-          <View style={styles.emailPreview}>
-            <Text style={styles.emailPreviewTitle}>Email Preview</Text>
-            <Text style={styles.emailPreviewCompact}>
-              <Text style={styles.emailPreviewLabel}>To:</Text>{" "}
-              ndricim@bilmenu.com
-              {"\n"}
-              <Text style={styles.emailPreviewLabel}>Subject:</Text> BilMenu
-              Image Submission
-              {"\n"}
-            </Text>
-
-            <View style={styles.subjectNameContainer}>
-              <Text
-                style={[
-                  styles.subjectNameText,
-                  userName.length > 0 && styles.subjectNameTextBold,
-                ]}
-              >
-                By: {userName || "No name provided"}
-              </Text>
-            </View>
-
-            <Text style={styles.emailPreviewCompact}>
-              <Text style={styles.emailPreviewLabel}>Message:</Text> Hi! 👋 This
-              meal was missing from BilMenu, so I'm sharing a photo. This is:{" "}
-              {selectedMeal}. Thanks for BilMenu!
-            </Text>
-          </View>
-        </View>
-
-        <View style={styles.nameInputContainer}>
-          <View style={styles.nameInputHeader}>
-            <View style={styles.nameInputLabelContainer}>
-              <Text style={styles.nameInputLabel}>
-                Add your name for attribution{" "}
-                <Text style={styles.optionalText}>(optional)</Text>
-                <TouchableOpacity
-                  style={styles.helpButton}
-                  onPress={() => setShowAttributionModal(true)}
-                >
-                  <Text style={styles.helpButtonText}>?</Text>
-                </TouchableOpacity>
-              </Text>
-            </View>
-            <Text
-              style={[
-                styles.characterCounter,
-                userName.length === 50 && styles.characterCounterLimit,
-              ]}
-            >
-              {userName.length}/50
-            </Text>
-          </View>
-          <View style={styles.nameInputWrapper}>
-            <TextInput
-              style={styles.nameInput}
-              placeholder="Your name"
-              value={userName}
-              onChangeText={setUserName}
-              placeholderTextColor={BilMenuTheme.colors.textLight}
-              maxLength={50}
-            />
-            {userName.length > 0 && (
-              <TouchableOpacity
-                style={styles.clearButton}
-                onPress={() => setUserName("")}
-              >
-                <Text style={styles.clearButtonText}>×</Text>
-              </TouchableOpacity>
-            )}
-          </View>
-        </View>
-
-        <TouchableOpacity style={styles.sendButton} onPress={handleSendEmail}>
-          <Text style={styles.sendButtonText}>📧 Send Email</Text>
-        </TouchableOpacity>
-      </ScrollView>
-
-      <TouchableOpacity style={styles.backButton} onPress={() => setStep(2)}>
-        <Text style={styles.backButtonText}>← Back to Camera</Text>
-      </TouchableOpacity>
-
-      <Modal
-        visible={showAttributionModal}
-        transparent={true}
-        animationType="fade"
-        onRequestClose={() => setShowAttributionModal(false)}
-      >
-        <View style={styles.modalOverlay}>
-          <View style={styles.modalContent}>
-            <Text style={styles.modalTitle}>Attribution</Text>
-            <Text style={styles.modalText}>
-              Your name will be shown as a thank you to contributors who helped
-              improve the meal image collection.
-            </Text>
-            <Text style={styles.modalText}>
-              It will appear in a contributors section to recognize your help in
-              making BilMenu better for all students.
-            </Text>
-            <TouchableOpacity
-              style={styles.modalButton}
-              onPress={() => setShowAttributionModal(false)}
-            >
-              <Text style={styles.modalButtonText}>Got it!</Text>
-            </TouchableOpacity>
-          </View>
-        </View>
-      </Modal>
-    </KeyboardAvoidingView>
+    <Step3
+      selectedMeal={selectedMeal}
+      capturedImage={capturedImage}
+      imageSize={imageSize}
+      userName={userName}
+      onUserNameChange={setUserName}
+      onSendEmail={handleSendEmail}
+      onBack={() => setStep(2)}
+      showAttributionModal={showAttributionModal}
+      onToggleAttributionModal={setShowAttributionModal}
+    />
   );
 
   return (
@@ -841,557 +246,5 @@ const styles = StyleSheet.create({
     textAlign: "center",
     marginBottom: 20,
     lineHeight: 24,
-  },
-  stepContainer: {
-    flex: 1,
-  },
-  step2Content: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  stepTitle: {
-    fontSize: 20,
-    fontWeight: "bold",
-    color: BilMenuTheme.colors.textWhite,
-    marginBottom: 20,
-    textAlign: "center",
-  },
-  searchInput: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: 10,
-    padding: 15,
-    fontSize: 16,
-    color: BilMenuTheme.colors.text,
-    marginBottom: 20,
-    borderWidth: 1,
-    borderColor: BilMenuTheme.colors.border,
-  },
-  mealsList: {
-    flex: 1,
-    borderWidth: 1,
-    borderColor: BilMenuTheme.colors.border,
-    borderRadius: 10,
-    padding: 10,
-    backgroundColor: BilMenuTheme.colors.surfaceLight,
-  },
-  mealButton: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: 8,
-    padding: 12,
-    marginBottom: 8,
-    borderWidth: 1,
-    borderColor: BilMenuTheme.colors.border,
-    flexDirection: "row",
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 1,
-    },
-    shadowOpacity: 0.1,
-    shadowRadius: 2,
-    elevation: 2,
-  },
-  mealButtonIcon: {
-    fontSize: 18,
-    color: BilMenuTheme.colors.secondary,
-    fontWeight: "bold",
-    marginRight: 8,
-  },
-  mealButtonText: {
-    fontSize: 14,
-    color: BilMenuTheme.colors.text,
-    flex: 1,
-    textAlign: "left",
-  },
-  loadingText: {
-    fontSize: 16,
-    color: BilMenuTheme.colors.textMuted,
-    textAlign: "center",
-    marginTop: 50,
-  },
-  noResultsText: {
-    fontSize: 16,
-    color: BilMenuTheme.colors.textMuted,
-    textAlign: "center",
-    marginTop: 50,
-  },
-  selectedMealText: {
-    fontSize: 18,
-    textAlign: "center",
-    marginBottom: 30,
-  },
-  selectedLabel: {
-    color: BilMenuTheme.colors.textMuted,
-    fontSize: 18,
-    fontWeight: "500",
-  },
-  selectedMealName: {
-    color: BilMenuTheme.colors.secondary,
-    fontSize: 18,
-    fontWeight: "600",
-  },
-  captureButton: {
-    backgroundColor: BilMenuTheme.colors.secondary,
-    borderRadius: 15,
-    padding: 20,
-    alignItems: "center",
-    marginBottom: 20,
-  },
-  captureButtonText: {
-    fontSize: 18,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "bold",
-  },
-  galleryButton: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: 15,
-    padding: 20,
-    alignItems: "center",
-    marginTop: 15,
-    borderWidth: 1,
-    borderColor: BilMenuTheme.colors.border,
-  },
-  galleryButtonText: {
-    fontSize: 18,
-    color: BilMenuTheme.colors.text,
-    fontWeight: "bold",
-  },
-  imagePreview: {
-    alignItems: "center",
-    marginBottom: 30,
-  },
-  previewImage: {
-    width: 200,
-    height: 150,
-    borderRadius: 10,
-    borderWidth: 2,
-    borderColor: BilMenuTheme.colors.border,
-  },
-  sendButton: {
-    backgroundColor: BilMenuTheme.colors.secondary,
-    borderRadius: 12,
-    padding: 15,
-    alignItems: "center",
-    marginBottom: 20,
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.2,
-    shadowRadius: 4,
-    elevation: 4,
-  },
-  sendButtonText: {
-    fontSize: 16,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "bold",
-  },
-  nameInputContainer: {
-    marginBottom: 15,
-  },
-  nameInputHeader: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    marginBottom: 8,
-  },
-  nameInputLabelContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    flex: 1,
-  },
-  nameInputLabel: {
-    fontSize: 14,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "500",
-    flex: 1,
-  },
-  helpButton: {
-    width: 16,
-    height: 16,
-    borderRadius: 8,
-    backgroundColor: BilMenuTheme.colors.secondary,
-    justifyContent: "center",
-    alignItems: "center",
-    marginLeft: 6,
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 1,
-    },
-    shadowOpacity: 0.2,
-    shadowRadius: 2,
-    elevation: 2,
-  },
-  optionalText: {
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "500",
-  },
-  helpButtonText: {
-    fontSize: 12,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "bold",
-  },
-  characterCounter: {
-    fontSize: 12,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "400",
-  },
-  characterCounterLimit: {
-    color: "#FF4444",
-    fontWeight: "600",
-  },
-  nameInputWrapper: {
-    position: "relative",
-  },
-  nameInput: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: 8,
-    padding: 12,
-    paddingRight: 40,
-    borderWidth: 1,
-    borderColor: BilMenuTheme.colors.border,
-    fontSize: 14,
-    color: BilMenuTheme.colors.text,
-  },
-  clearButton: {
-    position: "absolute",
-    right: 8,
-    top: "50%",
-    transform: [{ translateY: -12 }],
-    width: 24,
-    height: 24,
-    borderRadius: 12,
-    backgroundColor: BilMenuTheme.colors.textLight,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  clearButtonText: {
-    fontSize: 16,
-    color: BilMenuTheme.colors.surface,
-    fontWeight: "bold",
-    lineHeight: 16,
-  },
-  step3Content: {
-    flex: 1,
-    marginBottom: 20,
-  },
-  step3ScrollContent: {
-    flexGrow: 1,
-    paddingBottom: 20,
-  },
-  previewContainer: {
-    flexDirection: "row",
-    alignItems: "flex-start",
-    marginBottom: 15,
-    gap: 12,
-  },
-  imagePreviewContainer: {
-    alignItems: "center",
-  },
-  imagePreviewSmall: {
-    width: 100,
-    height: 100,
-    borderRadius: 8,
-    borderWidth: 2,
-    borderColor: BilMenuTheme.colors.border,
-    overflow: "hidden",
-  },
-  previewImageSmall: {
-    width: "100%",
-    height: "100%",
-    resizeMode: "cover",
-  },
-  imageSizeText: {
-    fontSize: 10,
-    color: BilMenuTheme.colors.textWhite,
-    marginTop: 4,
-    textAlign: "center",
-  },
-  imageFilenameText: {
-    fontSize: 9,
-    color: BilMenuTheme.colors.textWhite,
-    marginTop: 2,
-    textAlign: "center",
-    fontStyle: "italic",
-  },
-  emailPreview: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: 8,
-    padding: 8,
-    flex: 1,
-    borderWidth: 1,
-    borderColor: BilMenuTheme.colors.border,
-  },
-  emailPreviewTitle: {
-    fontSize: 12,
-    fontWeight: "bold",
-    color: BilMenuTheme.colors.text,
-    marginBottom: 6,
-  },
-  emailPreviewCompact: {
-    fontSize: 11,
-    color: BilMenuTheme.colors.text,
-    lineHeight: 14,
-  },
-  emailPreviewLabel: {
-    fontSize: 11,
-    fontWeight: "600",
-    color: BilMenuTheme.colors.text,
-  },
-  subjectNameContainer: {
-    height: 32,
-    justifyContent: "center",
-    marginTop: 1,
-    marginBottom: 1,
-  },
-  subjectNameText: {
-    fontSize: 11,
-    color: BilMenuTheme.colors.text,
-    fontWeight: "500",
-  },
-  subjectNameTextBold: {
-    fontWeight: "bold",
-  },
-  backButton: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: 10,
-    padding: 15,
-    alignItems: "center",
-    borderWidth: 1,
-    borderColor: BilMenuTheme.colors.border,
-    marginTop: "auto",
-    marginBottom: 20,
-  },
-  backButtonText: {
-    fontSize: 16,
-    color: BilMenuTheme.colors.text,
-  },
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: "rgba(0, 0, 0, 0.5)",
-    justifyContent: "center",
-    alignItems: "center",
-    padding: 20,
-  },
-  modalContent: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: 12,
-    padding: 20,
-    maxWidth: 300,
-    width: "100%",
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: "bold",
-    color: BilMenuTheme.colors.text,
-    marginBottom: 12,
-    textAlign: "center",
-  },
-  modalText: {
-    fontSize: 14,
-    color: BilMenuTheme.colors.text,
-    lineHeight: 20,
-    marginBottom: 12,
-  },
-  modalButton: {
-    backgroundColor: BilMenuTheme.colors.secondary,
-    borderRadius: 8,
-    padding: 12,
-    alignItems: "center",
-    marginTop: 8,
-  },
-  modalButtonText: {
-    fontSize: 16,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "bold",
-  },
-  weekInfo: {
-    marginBottom: 20,
-    alignItems: "center",
-  },
-  weekInfoText: {
-    fontSize: 16,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "600",
-    marginBottom: 5,
-  },
-  lastUpdatedText: {
-    fontSize: 12,
-    color: BilMenuTheme.colors.textMuted,
-    marginBottom: 3,
-  },
-  checkpointCountText: {
-    fontSize: 12,
-    color: BilMenuTheme.colors.textMuted,
-    marginBottom: 10,
-  },
-  toggleContainer: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: BilMenuTheme.borderRadius.medium,
-    padding: BilMenuTheme.spacing.md,
-    marginBottom: 20,
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.1,
-    shadowRadius: 8,
-    elevation: 4,
-  },
-  toggleRow: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    paddingVertical: BilMenuTheme.spacing.xs,
-  },
-  toggleInfo: {
-    flex: 1,
-    marginRight: BilMenuTheme.spacing.sm,
-  },
-  toggleLabel: {
-    fontSize: BilMenuTheme.typography.body.fontSize,
-    fontWeight: BilMenuTheme.typography.subtitle.fontWeight,
-    color: BilMenuTheme.colors.text,
-  },
-  toggleDescription: {
-    fontSize: BilMenuTheme.typography.small.fontSize,
-    color: BilMenuTheme.colors.textLight,
-    lineHeight: BilMenuTheme.typography.small.lineHeight,
-    marginTop: 2,
-  },
-  resultsCount: {
-    fontSize: 12,
-    color: BilMenuTheme.colors.textMuted,
-    textAlign: "center",
-    marginBottom: 12,
-    fontStyle: "italic",
-  },
-  viewSelector: {
-    flexDirection: "row",
-    marginBottom: 20,
-    backgroundColor: BilMenuTheme.colors.surfaceLight,
-    borderRadius: 10,
-    padding: 4,
-  },
-  viewButton: {
-    flex: 1,
-    paddingVertical: 12,
-    paddingHorizontal: 16,
-    borderRadius: 8,
-    alignItems: "center",
-  },
-  viewButtonActive: {
-    backgroundColor: BilMenuTheme.colors.surface,
-  },
-  viewButtonText: {
-    fontSize: 14,
-    color: BilMenuTheme.colors.textMuted,
-    fontWeight: "500",
-  },
-  viewButtonTextActive: {
-    color: BilMenuTheme.colors.text,
-    fontWeight: "600",
-  },
-  sectionTitle: {
-    fontSize: 16,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "600",
-    marginBottom: 10,
-  },
-  daysContainer: {
-    marginBottom: 20,
-  },
-  daysGrid: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    justifyContent: "space-between",
-  },
-  dayButton: {
-    backgroundColor: BilMenuTheme.colors.surface,
-    borderRadius: 6,
-    paddingVertical: 2,
-    paddingHorizontal: 6,
-    marginBottom: 8,
-    alignItems: "center",
-    justifyContent: "center",
-    flexBasis: "13.5%",
-    minWidth: 48,
-    height: 32,
-    borderWidth: 1,
-    borderColor: BilMenuTheme.colors.border,
-  },
-  dayButtonActive: {
-    backgroundColor: BilMenuTheme.colors.secondary,
-  },
-  dayButtonToday: {
-    borderColor: BilMenuTheme.colors.secondary,
-    borderWidth: 2,
-  },
-  dayButtonText: {
-    fontSize: 11,
-    color: BilMenuTheme.colors.text,
-    fontWeight: "500",
-  },
-  dayButtonTextActive: {
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "600",
-  },
-  todayLabel: {
-    fontSize: 8,
-    color: BilMenuTheme.colors.text,
-    fontWeight: "600",
-    marginTop: 1,
-  },
-  todayLabelActive: {
-    color: BilMenuTheme.colors.textWhite,
-  },
-  mealTypeContainer: {
-    marginBottom: 20,
-  },
-  mealTypeButtons: {
-    flexDirection: "row",
-    backgroundColor: "rgba(255, 255, 255, 0.1)",
-    borderRadius: 10,
-    borderWidth: 1,
-    borderColor: "rgba(255, 255, 255, 0.2)",
-    overflow: "hidden",
-  },
-  mealTypeButton: {
-    flex: 1,
-    paddingVertical: 12,
-    paddingHorizontal: 8,
-    alignItems: "center",
-    borderRightWidth: 1,
-    borderRightColor: "rgba(255, 255, 255, 0.2)",
-  },
-  mealTypeButtonLeft: {
-    borderTopLeftRadius: 9,
-    borderBottomLeftRadius: 9,
-  },
-  mealTypeButtonCenter: {
-    // No additional border radius for center
-  },
-  mealTypeButtonRight: {
-    borderTopRightRadius: 9,
-    borderBottomRightRadius: 9,
-    borderRightWidth: 0, // Remove right border for last button
-  },
-  mealTypeButtonActive: {
-    backgroundColor: "#FF9434",
-  },
-  mealTypeButtonText: {
-    fontSize: 12,
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "500",
-    textAlign: "center",
-  },
-  mealTypeButtonTextActive: {
-    color: BilMenuTheme.colors.textWhite,
-    fontWeight: "600",
   },
 });
