+++ mobile/bilmenu/components/submit/index.ts
@@ -1,0 +1,5 @@
+export { Step1 } from "./Step1";
+export { Step2 } from "./Step2";
+export { Step3 } from "./Step3";
+export { DayMealView } from "./DayMealView";
+export { SearchView } from "./SearchView";
